requirements.txt
langchain-community
python-dotenv
neo4j